requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.4
python-socketio==5.9.0
python-engineio==4.7.1

torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
diffusers==0.21.4
transformers==4.32.1
accelerate==0.22.0
safetensors==0.3.1

google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
google-api-python-client==2.99.0

pillow==10.0.0
opencv-python==4.8.1.78
numpy==1.24.3
scipy==1.11.3

einops==0.7.0
omegaconf==2.3.0
xformers==0.0.22
controlnet-aux==0.0.7

pydantic==2.3.0
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1

# Model downloading
huggingface-hub==0.17.3

# Jupyter notebook support (for Colab)
ipywidgets==8.1.1
jupyter==1.0.0
